[join] update rules for which fields to copy
Before, when a the fields= option was missing, all fields in the source
table were copied to the target data table. This is a sensible default for
one-to-one joins, but makes less sense for many-to-one joins.

The new calc= option is designed for many-to-one joins. With this commit,
if the calc= option is present and files= is missing, mapshaper does not
copy all fields from the source table by default. If this is what you want,
you have to specify fields=*

The best way to handle many-to-one joins is using the new calc= option, which
makes the user specify how each source record should be